Create 🦉 Night Owl
People tend to be either early birds or night owls. There are several factors that influence this, such as genetics, age, lifestyle, and more.

Let's use conditionals to see which one you are!

Start by creating a wakeUpvariable and set it equal to the time you wake up, between 0 and 23.

Test to see if you're an early bird or a night owl with an if/ :else statement

If wakeUpes < 8, log "Early bird gets the worm! 🌅"
Else, log "But who wants the worm? 🌃"